More progress, wired threat range to check for "Keen" and calculate accordingly.

+++ trunk/ItemSmithWorkShop/ItemSmithWorkShop/Items/Interfaces/IWeaponEnchantment.cs
@@ -9,7 +9,7 @@ namespace ItemSmithWorkShop.Items.Interfaces
 	{
 		string EnchantmentName { get; }
 		string Affix { get; }
-		int CostModifier { get; }
+		double CostModifier { get; }
 		string WeaponUse { get; }
 		string StandardDamageBonus { get; }
 		bool CriticalDamageBonus { get; }
